+++ src/yetibot/core/repl.clj
@@ -15,6 +15,7 @@
 (defn load-minimal []
   (require 'yetibot.core.commands.echo :reload)
   (require 'yetibot.core.commands.help :reload)
+  (require 'yetibot.core.commands.default-command :reload)
   (require 'yetibot.core.commands.room :reload)
   (require 'yetibot.core.observers.history :reload)
   (require 'yetibot.core.commands.history :reload)